leaving with virtual sessions

+++ modules/data/src/main/kotlin/agartha/data/services/IPractitionerService.kt
@@ -89,5 +89,5 @@ interface IPractitionerService : IBaseService<PractitionerDBO> {
      * @return true if practitioner successfully paid the contributionsPoints
      */
     fun payForAddingVirtualSessions(
-            practitioner: PractitionerDBO, numberOfSessions: Long): Boolean
+            practitioner: PractitionerDBO, numberOfSessions: Int): Boolean
 }
